Add htmx for cart app views (#16)
- Add htmx for clear cart,  update cart and remove item from cart views.

+++ store/cart/views.py
@@ -1,12 +1,15 @@
+from django.contrib.auth.decorators import login_required
 from django.core.exceptions import ObjectDoesNotExist
-from django.http import HttpRequest, HttpResponse, HttpResponseRedirect
+from django.http import HttpRequest, HttpResponse
 from django.shortcuts import redirect, render
+from django.views.decorators.http import require_POST
 
 from store.products.models import Product
 
 from .models import Cart, CartItem
 
 
+@login_required
 def cart_summary_view(request: HttpRequest) -> HttpResponse:
     """A view to show cart summary.
 
@@ -25,6 +28,7 @@ def cart_summary_view(request: HttpRequest) -> HttpResponse:
     return render(request, "store/cart_summary.html", {"cart": cart})
 
 
+@login_required
 def add_to_cart(request: HttpRequest, product_id: str) -> HttpResponse:
     """A view to add a product to cart.
 
@@ -42,8 +46,9 @@ def add_to_cart(request: HttpRequest, product_id: str) -> HttpResponse:
     except ObjectDoesNotExist:
         cart = Cart.objects.create(user=request.user)
 
+    product = Product.objects.get(id=product_id)
+
     if request.method == "POST":
-        product = Product.objects.get(id=product_id)
         if product.stocks > 0:
             try:
                 cart_item = CartItem.objects.get(cart=cart, product=product)
@@ -54,9 +59,11 @@ def add_to_cart(request: HttpRequest, product_id: str) -> HttpResponse:
             cart_item.quantity += int(quantities)
             cart_item.save()
 
-    return HttpResponseRedirect(request.META.get("HTTP_REFERER"))
+    return redirect("products:product_detail", slug=product.slug)
 
 
+@login_required
+@require_POST
 def remove_from_cart_view(request: HttpRequest, product_id: str) -> HttpResponse:
     """A view to remove a product from cart.
 
@@ -66,19 +73,23 @@ def remove_from_cart_view(request: HttpRequest, product_id: str) -> HttpResponse
     Returns:
         HttpResponse: The response object.
     """
+    if request.htmx:
+        try:
+            if request.user.cart:
+                cart = request.user.cart
+        except ObjectDoesNotExist:
+            cart = Cart.objects.create(user=request.user)
 
-    try:
-        if request.user.cart:
-            cart = request.user.cart
-    except ObjectDoesNotExist:
-        cart = Cart.objects.create(user=request.user)
+        cart_item = CartItem.objects.get(cart=cart, product_id=product_id)
+        if cart_item:
+            cart_item.delete()
+        return render(request, "_partials/cart_update.html", {"cart": cart})
 
-    cart_item = CartItem.objects.get(cart=cart, product_id=product_id)
-    if cart_item:
-        cart_item.delete()
     return redirect("cart:summary")
 
 
+@login_required
+@require_POST
 def clear_cart_view(request: HttpRequest) -> HttpResponse:
     """A view to clear cart.
 
@@ -89,7 +100,7 @@ def clear_cart_view(request: HttpRequest) -> HttpResponse:
         HttpResponse: The response object.
     """
 
-    if request.method == "POST":
+    if request.htmx:
         try:
             if request.user.cart:
                 cart = request.user.cart
@@ -97,11 +108,13 @@ def clear_cart_view(request: HttpRequest) -> HttpResponse:
             cart = Cart.objects.create(user=request.user)
 
         cart.cart_items.all().delete()
-        return redirect("cart:summary")
-    else:
-        return HttpResponseRedirect(request.META.get("HTTP_REFERER"))
+        return render(request, "_partials/clear_cart.html")
+
+    return redirect("cart:summary")
 
 
+@require_POST
+@login_required
 def update_cart_view(request: HttpRequest) -> HttpResponse:
     """A view to update cart.
 
@@ -112,7 +125,7 @@ def update_cart_view(request: HttpRequest) -> HttpResponse:
         HttpResponse: The response object.
     """
 
-    if request.method == "POST":
+    if request.htmx:
         try:
             if request.user.cart:
                 cart = request.user.cart
@@ -125,6 +138,6 @@ def update_cart_view(request: HttpRequest) -> HttpResponse:
             cart_item = CartItem.objects.get(cart=cart, product_id=product_id)
             cart_item.quantity = quantity
             cart_item.save()
-        return redirect("cart:summary")
-    else:
-        return HttpResponseRedirect(request.META.get("HTTP_REFERER"))
+        return render(request, "_partials/cart_update.html", {"cart": cart})
+
+    return redirect("cart:summary")
